Update export status in spark job rather than in S3
