+++ app-backend/db/src/main/scala/com/azavea/rf/database/ExportDao.scala
@@ -23,10 +23,6 @@ import java.sql.Timestamp
 import java.util.{Date, UUID}
 import java.net.URI
 
-import cats.free.Free
-import doobie.free.connection
-
-
 object ExportDao extends Dao[Export] {
 
   val tableName = "exports"
@@ -38,6 +34,12 @@ object ExportDao extends Dao[Export] {
       visibility, toolrun_id, export_options
     FROM
   """ ++ tableF
+
+  def unsafeGetExportById(exportId: UUID): ConnectionIO[Export] =
+    query.filter(exportId).select
+
+  def getExportById(exportId: UUID): ConnectionIO[Option[Export]] =
+    query.filter(exportId).selectOption
 
   def insert(export: Export, user: User): ConnectionIO[Export] = {
     val insertF: Fragment = Fragment.const(s"INSERT INTO ${tableName} (")
@@ -64,7 +66,7 @@ object ExportDao extends Dao[Export] {
         modified_by = ${user.id},
         export_status = ${export.exportStatus},
         visibility = ${export.visibility}
-      WHERE id = ${id} AND owner = ${user.id}
+      WHERE id = ${id}
     """).update.run
   }
